+++ support/cas-server-support-jdbc-monitor/src/main/java/org/apereo/cas/monitor/config/CasJdbcMonitorConfiguration.java
@@ -37,7 +37,7 @@ public class CasJdbcMonitorConfiguration {
             final ExecutorService executor) {
         final JdbcDataSourceMonitor m = new JdbcDataSourceMonitor(monitorDataSource());
         m.setValidationQuery(casProperties.getMonitor().getJdbc().getValidationQuery());
-        m.setMaxWait(casProperties.getMonitor().getJdbc().getMaxWait());
+        m.setMaxWait(Long.valueOf(casProperties.getMonitor().getJdbc().getMaxWait()).intValue());
         m.setExecutor(executor);
         return m;
     }
